% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_tvem_parallel.R
\name{fit_tvem}
\alias{fit_tvem}
\title{Fit tvem model(s)}
\usage{
fit_tvem(
  tvem_outcome_variable,
  tvem_time_variable,
  tvem_id_variable,
  tvem_variate_effects = NULL,
  tvem_invariate_effects,
  tvem_data,
  tvem_knots,
  fit_permuted = FALSE,
  permutation_sets = NULL
)
}
\description{
fit_tvem is a wrapper for the 'tvem' function in the 'tvem' package.
}
\details{
@details The primary utility of fit_tvem is to accept lists of data objects and return a list of fitted tvem models.

@param tvem_outcome_variable A string of length 1 which denotes the name of the dependent variable.
@param tvem_time_variable A sting of length 1 which denotes the name of the variable which represents time (e.g., age or minutes).
@param tvem_id_variable A string of length 1 which denotes the name of the variable which represents each particpant.
@param tvem_variate_effects A string of any length (which is reasonable to model) which denotes the names of the variables which should be modeled as time-varying effects.
@param tvem_invariate_effects A string of any length (which is reasonable to model) which denotes theh names of the variables which should be modeled as time-invariant effects.
@param tvem_data Either a dataframe or a list of dataframes
@param tvem_knots The number of knots to use in the tvem model. For an explination of knots, see the documentation for the 'tvem' package.

@returns Returns the output from the 'tvem' function (which is a list containing information about model fit). This function returns the tvem output itself (if the inputed dataset is a dataframe) or a list of tvem output (if the inputed data was a list of dataframes) fit for each inputed dataframe.
}
