% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_fit.R
\name{median_fit}
\alias{median_fit}
\title{Extract median fit statistics from across multiple imputations}
\usage{
median_fit(model.fit.list)
}
\description{
Extract median fit statistics from across multiple imputations
}
