% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_outcome.R
\name{permute_outcome}
\alias{permute_outcome}
\title{Permute outcome variable}
\usage{
permute_outcome(variable.to.permute, n.perm)
}
\arguments{
\item{n.perm}{An integer that designates the number of permuted datasets to create.}

\item{observed.model.fit}{A vector of values to be imputed.}
}
\value{
A matrix of y*x, where y is the length of the outcome variable vector and x is equal to the number of permutations. Each permuted vector of the output variable is stored as a column in the matrix.
}
\description{
Generates permutations.
}
\details{
Accepts a tvem model or a list of tvem models, extracts the outcome variable from the first model, and returns the desired number of permuted sets of the outcome variable.
}
