% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_invariate.R
\name{pool.invariate.estimates}
\alias{pool.invariate.estimates}
\title{Pool invariant coefficient estimates}
\usage{
pool.invariate.estimates(
  model.fit.list,
  effect.path = "invar_effects_estimates",
  return.term = "tstat"
)
}
\arguments{
\item{model.fit.list}{List composed of elements which are 'tvem' objects. Each 'tvem' object should be a model fit using the tvem package on one iteration of a multiply imputed dataset.}

\item{effect.path}{A string that denotes the subelement of each 'tvem' object where the invariant effects are located. Defaults to 'invar_effects_estimates', which is the default string used by 'tvem' objects.}

\item{return.term}{A string that determines if t-statistics or beta coefficients and standard errors are returned. Accepts 'tstat' or 'beta' as inputs.}
}
\value{
Returns a dataframe composed of either t-statistics or beta coefficients and standard errors.
}
\description{
Accepts a list of arbitrary length, for which each list element contains a tvem model fit on a multiply imputed dataset. Returns pooled statistics.
}
