% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_variate.R
\name{pool.variate.estimates}
\alias{pool.variate.estimates}
\title{Pool variant coefficient estimates}
\usage{
pool.variate.estimates(
  model.fit.list,
  effect.path = "grid_fitted_coefficients",
  return.term = "tstat"
)
}
\arguments{
\item{model.fit.list}{List composed of elements which are 'tvem' objects. Each 'tvem' object should be a model fit using the tvem package on one iteration of a multiply imputed dataset.}

\item{effect.path}{A string that denotes the subelement of each 'tvem' object where the variant effects are located. Defaults to 'grid_fitted_coefficients', which is the default string used by 'tvem' objects.}

\item{return.term}{A string that determines if t-statistics or beta coefficients and standard errors are returned. Accepts 'tstat' or 'beta' as inputs.}
}
\value{
Returns a list of dataframes (one list element per variant effect) composed of either t-statistics or beta coefficients and standard errors (one value for each point on the time grid).
}
\description{
Accepts a list of arbitrary length, for which each list element contains a tvem model fit on a multiply imputed dataset. Returns pooled statistics.
}
