% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_tvem.R
\name{pool_tvem}
\alias{pool_tvem}
\title{Pool coefficients from \code{tvem} models}
\usage{
pool_tvem(obs.model, perm.model)
}
\arguments{
\item{obs.model}{List of \code{tvem} models fit on multiply imputed data.}

\item{perm.model}{List of \code{tvem} models fit on permuted data.}
}
\value{
A list with a length of two. First list element contains the beta and p-values for invariant effects. Second list element contains the beta and p-values for the variate effects.
}
\description{
Utility function used to pool coefficients from \code{tvem} models fit on multiply imputed data.
}
\details{
Uses the \code{pool.variate.estimates} and \code{pool.invariate.estimates} functions to pool beta values and generate permuted p-values across multiply imputed datasets.
}
