% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool_wald.R
\name{pool_wald}
\alias{pool_wald}
\title{Pool multivariate Wald test statistic}
\usage{
pool_wald(model1, model0)
}
\arguments{
\item{model1}{The model which has a larger number of variables.}

\item{model0}{The model which has a smaller number of variables.}
}
\value{
Multivariate Wald statistic.
}
\description{
Calculate a pooled multivariate Wald statistic for a set of models fit on multiply imputed data sets.
}
\references{
Li, K. H., T. E. Raghunathan, and D. B. Rubin. 1991.
Large-Sample Significance Levels from Multiply Imputed Data Using
Moment-Based Statistics and an F Reference Distribution.
\emph{Journal of the American Statistical Association}, 86(416): 1065–73.
}
