% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvem_model_comparison.R
\name{tvem_model_comparison}
\alias{tvem_model_comparison}
\title{Compare nested tvem models}
\usage{
tvem_model_comparison(full_model, reduced_model)
}
\arguments{
\item{full_model}{A list of length two. The first list element is assumed to be the output from the observed model. The second list element is assumed to be the output from the permuted models. These models should include more parameters than the reduced (AKA - the null) model.}

\item{reduced_model}{A list of length two. The first list element is assumed to be the output from the observed model. The second list element is assumed to be the output from the permuted models. These models should include fewer parameters than the full model.}
}
\value{
Returns a list of length two. The first list element will be the pooled invariate and variate statistics with their associated permuted p-values. The second list element will be the multivariate wald statistic for the nested models and its associated permuted p-value.
}
\description{
tvem_model_comparison pools coefficients and test statistics.
}
\details{
tvem_model_comparison pools coefficients and test statistics across imputations to allow for nested model comparisons and the derviation of permutation p-values.
}
